refactor(PersistenceManager):Refatoração dos metodos de carregamento dos dados

+++ src/main/java/br/edu/ifba/models/Emprestimo.java
@@ -45,7 +45,7 @@ public class Emprestimo {
         this.atrasado = false; // Inicia em dia
     }
 
-    public Emprestimo(int id,
+    public Emprestimo(long id,
                       Usuario usuario,
                       Livro livro,
                       LocalDate dataEmprestimo,
